+++ classes/Helper.php
@@ -17,13 +17,38 @@ class Helper
     const GIT_REGEX = '/(?:git|ssh|https?|git@[-\w.]+):(\/\/)?(.*?)(\.git)(\/?|\#[-\d\w._]+?)$/';
 
     /**
+     * Checks if git-sync is properly configured with a repository URL
+     *
+     * @return bool
+     */
+    public static function isGitSyncConfigured()
+    {
+        $config = Grav::instance()['config']->get('plugins.git-sync');
+        $repository = $config['repository'] ?? null;
+        return !empty($repository);
+    }
+
+    /**
+     * Checks if git-sync is ready to use (installed, configured, and initialized)
+     *
+     * @return bool
+     */
+    public static function isGitSyncReady()
+    {
+        return static::isGitInstalled() && static::isGitSyncConfigured() && static::isGitInitialized();
+    }
+
+    /**
      * Checks if the user/ folder is already initialized
      *
      * @return bool
      */
     public static function isGitInitialized()
     {
-        return file_exists(rtrim(USER_DIR, '/') . '/.git');
+        /** @var Config $grav */
+        $config = Grav::instance()['config']->get('plugins.git-sync');
+        $repositoryPath = isset($config['local_repository']) && $config['local_repository'] ? $config['local_repository'] : USER_DIR;
+        return file_exists(rtrim($repositoryPath, '/') . '/.git');
     }
 
     /**
@@ -77,7 +102,7 @@ class Helper
             return $repository;
         }
 
-        return str_replace('://', "://${user}${password}@", $repository);
+        return str_replace('://', "://{$user}{$password}@", $repository);
     }
 
     /**