+++ cli/SyncCommand.php
@@ -46,6 +46,8 @@ class SyncCommand extends ConsoleCommand
         $plugin->sync();
 
         $this->output->writeln('completed.');
+
+        return 0;
     }
 }